+++ terminal_games/snake.c
@@ -190,7 +190,7 @@ int randn(int limit)
 
 {
 
-    return (long)limit*rand()/RAND_MAX;
+    return (int)limit*rand()/RAND_MAX;
 
 }
